create version specific Natural types
