+++ core/shared/src/main/scala/zio/prelude/newtypes/package.scala
@@ -17,6 +17,8 @@
 package zio.prelude
 
 package object newtypes {
+  type Natural = Natural.Type
+
   object Sum extends SubtypeF
 
   /**
@@ -112,35 +114,4 @@ package object newtypes {
 
   type FailureOut[+A] = FailureOut.Type[A]
 
-  object Natural extends Subtype[Int]() {
-
-    val one: Natural =
-      Natural(1)
-
-    val zero: Natural =
-      Natural(0)
-
-    def successor(n: Natural): Natural =
-      wrap(n + 1)
-
-    def times(x: Natural, y: Natural): Natural = {
-      val product = x * y
-      if (x == 0 || product / x != y) Natural(Int.MaxValue) else wrap(product)
-    }
-
-    def plus(x: Natural, y: Natural): Natural = {
-      val sum = x + y
-      if (sum < 0) Natural(Int.MaxValue) else wrap(sum)
-    }
-
-    def minus(x: Natural, y: Natural): Natural = {
-      val difference = x - y
-      if (difference < 0) zero else wrap(difference)
-    }
-
-    private[prelude] def unsafeMake(n: Int): Natural =
-      wrap(n)
-  }
-
-  type Natural = Natural.Type
 }
